requirements.txt
transformers==4.46.1
huggingface-hub==0.26.2
python-dotenv==1.0.1